requirements.txt
pillow==9.5.0
speechrecognition==3.10.0
psutil==5.9.4
wikipedia==1.4.0
deep_translator==1.10.1
langdetect==1.0.9
geocoder==1.38.1
feedparser==6.0.10
speedtest-cli==2.1.3
pyttsx3==2.90
opencv-python==4.7.0.72
pyaudio==0.2.13
pycaw==20230322
pync==2.0.3
pywhatkit==5.4
playsound==1.3.0